backend/requirements.txt
redis==4.3.4
backtrader
IbPy2==0.8.0
retrying==1.3.4
loguru==0.7.2
pandas
matplotlib
python-dotenv
fastapi 
uvicorn
yfinance